refactor: move templates package to node-app
