+++ imports/node-app/plugins/templates/i18n/de.json
