+++ imports/node-app/plugins/templates/i18n/en.json
