+++ imports/node-app/plugins/templates/i18n/fr.json
