+++ imports/node-app/plugins/templates/i18n/he.json
