+++ imports/node-app/plugins/templates/i18n/nb.json
