+++ imports/node-app/plugins/templates/i18n/sl.json
