+++ imports/node-app/registerPlugins.js
@@ -1,5 +1,6 @@
 /* node-app imports */
 import registerSystemInfoPlugin from "./plugins/system-info/index.js";
+import registerTemplatesPlugin from "./plugins/templates/index.js";
 
 /* meteor-app imports */
 import registerAccountsPlugin from "/imports/plugins/core/accounts/server/no-meteor/register";
@@ -37,7 +38,6 @@ import registerSurchargesPlugin from "/imports/plugins/included/surcharges/serve
 import registerTagsPlugin from "/imports/plugins/core/tags/server/no-meteor/register";
 import registerTaxesPlugin from "/imports/plugins/core/taxes/server/no-meteor/register";
 import registerTaxesRatesPlugin from "/imports/plugins/included/taxes-rates/server/no-meteor/register";
-import registerTemplatesPlugin from "/imports/plugins/core/templates/server/no-meteor/register";
 import registerTestAddressValidationPlugin from "/imports/plugins/included/address-validation-test/server/register";
 import registerUIPlugin from "/imports/plugins/core/ui/server/no-meteor/register";
