+++ imports/plugins/core/templates/server/simpleSchemas.js
